objectutils stingify recursion failsafe
ObjectUtilities.stringify now has a recursion detection build into it where it will detect if a key of a class is already used before in the same recursion flow and quits the loop by replacing the value with "[recursion overflow]".
with this comes a minor bug where an object that has un-enumerable properties places its brackets on a new line instead of after the key.