Update arrayUtilities.ahk
Added 1 level of depth the the check at the array utils. there were some instances where the recursion was already there but hiding within the first array, checking inside each array solved it, it doesnt go further then 1 depth level.

thats all it took... pretty disappointing if you ask me. i still dont think we are completely done with this tho so expect more frustration with this in the future.